Tool tree cleanup
